bkup-init/do nothing if db exists; bkup-stash/ exec bkup-init at the beginning;

+++ bin/bkup-init.sh
@@ -3,9 +3,13 @@
 # Required table: content list
 #   content: fid(md5),name,mode,date,base64(gziped)
 #   list: id(date),host,dist,owner,dir,fid
+# option (-r) : reset data
+. func.msg
+[ -s ~/.var/bkup.sq3 -a "$1" != -r ] && exit
 bkup-exec <<EOF
 drop table if exists content;
 drop table if exists list;
 create table content (id primary key,mode,date,base64);
 create table list (fid,host,dist,owner,name,dir,primary key(fid,host,dist));
 EOF
+_warn "Bkup DB initialized"
